% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{siverse_deps}
\alias{siverse_deps}
\title{List all siverse dependencies}
\usage{
siverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
siverse packages.}
}
\description{
List all siverse dependencies
}
