% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{siverse_packages}
\alias{siverse_packages}
\title{List all packages in the siverse}
\usage{
siverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include siverse in the list?}
}
\description{
List all packages in the siverse
}
\examples{
siverse_packages()
}
